SWAT-DEV Moved unused files to junk folder
