+++ src/com/meteorology/swat/junk/EncryptionAndHash.java
@@ -1,4 +1,4 @@
-package com.meteorology.swat.util;
+package com.meteorology.swat.junk;
 
 import java.sql.SQLException;
