+++ src/com/meteorology/swat/junk/MVCConfiguration.java
@@ -1,3 +1,4 @@
+package com.meteorology.swat.junk;
 /*package com.meteorology.swat.util;
 
 import javax.sql.DataSource;
