+++ src/com/meteorology/swat/junk/Normalize.java
@@ -1,3 +1,4 @@
+package com.meteorology.swat.junk;
 /*package com.meteorology.swat.util;
 
 import java.util.ArrayList;
